Examples move and docs refactor
