+++ packages/interchain/interchain-daemon/Cargo.toml
@@ -43,3 +43,5 @@ cosmwasm-std = "1.3.3"
 
 pretty_env_logger = "0.5.0"
 anyhow = { workspace = true }
+cw-orch-interchain = { workspace = true, features = ["daemon"] }
+cw-orch = { workspace = true, features = ["daemon"] }
